requirements.txt
numpy==1.18.5
tensorflow-gpu==2.3.0
opencv-python==4.4.0.42